dtgen: minor bugfix
There's a minor buffer-underrun issue in askyn() in datatypes/dtgen.c,
where the 'in' array is 8 chars long, but the fgets call says it can
read 127 characters.

This is only a problem if people write more than a simple 'y' or 'n',
so it's very unlikely that anyone ever hit it.